refactor(agents): move every adapter into its own directory
Restructure `src/agents/` so each adapter owns a folder containing its
production modules and tests. Same content, new layout:

  src/agents/claude/{index,plugins,plugin-apply,sanitize}.ts
  src/agents/claude/__tests__/{index,plugins,sanitize}.test.ts
  src/agents/codex/{index.ts,__tests__/index.test.ts}
  src/agents/copilot/{index.ts,__tests__/index.test.ts}
  src/agents/cursor/{index.ts,__tests__/index.test.ts}
  src/agents/vscode/{index.ts,__tests__/index.test.ts}

Imports inside each moved file rewritten for the new depth
(`../config` → `../../config`, `./_utils` → `../_utils`, etc.). The
public path `from "../agents/<name>"` keeps working because Node ESM
resolves a directory to its `index.ts`, so `src/agents/registry.ts`
and `src/migrate/migrate.ts` need no changes.

This was the original C1 shape from the architecture plan; the prior
commit landed it as sibling files (`claude-plugin-apply.ts` etc.) for
speed. Per-folder ownership is the durable structure — adding a new
adapter file or test now stays inside that adapter's directory.
