+++ src/agents/claude/index.ts
@@ -2,11 +2,11 @@ import { mkdir, readdir, readFile } from "node:fs/promises";
 import { homedir } from "node:os";
 import { join } from "node:path";
 import { log } from "@clack/prompts";
-import { AgentPaths } from "../config/paths";
-import type { AgentSyncConfig } from "../config/schema";
-import { denormalizeFromVault, denormalizeStringFromVault } from "../core/path-portability";
-import { sanitizeAndNormalizeJson, shouldNeverSync } from "../core/sanitizer";
-import { extractArchive } from "../core/tar";
+import { AgentPaths } from "../../config/paths";
+import type { AgentSyncConfig } from "../../config/schema";
+import { denormalizeFromVault, denormalizeStringFromVault } from "../../core/path-portability";
+import { sanitizeAndNormalizeJson, shouldNeverSync } from "../../core/sanitizer";
+import { extractArchive } from "../../core/tar";
 import {
   atomicWrite,
   collect,
@@ -14,16 +14,16 @@ import {
   readIfExists,
   type SnapshotArtifact,
   type SnapshotResult,
-} from "./_utils";
-import { applyClaudePluginsDir } from "./claude-plugin-apply";
-import { collectClaudePlugins } from "./claude-plugins";
+} from "../_utils";
+import { collectSkillArtifacts, InvalidSkillNameError, validateSkillName } from "../skills-walker";
+import { applyClaudePluginsDir } from "./plugin-apply";
+import { collectClaudePlugins } from "./plugins";
 import {
   sanitizeClaudeHooks,
   sanitizeClaudeMcp,
   sanitizeClaudePluginManifest,
   sanitizeClaudePluginMcp,
-} from "./claude-sanitize";
-import { collectSkillArtifacts, InvalidSkillNameError, validateSkillName } from "./skills-walker";
+} from "./sanitize";
 
 /** Snapshot payload for the Claude adapter. */
 export type ClaudeSnapshotResult = SnapshotResult;
@@ -397,7 +397,7 @@ export async function applyClaudeMarketplace(content: string): Promise<void> {
 // ─── Apply (pull side) ────────────────────────────────────────────────────────
 
 import { basename } from "node:path";
-import { decryptString } from "../core/encryptor";
+import { decryptString } from "../../core/encryptor";
 
 /** Read encrypted files from a vault subdirectory, ignoring missing directories. */
 async function readAgeFiles(dir: string): Promise<{ name: string; fullPath: string }[]> {
